♻️ Refacto du code pour la section AI
