+++ frontend/components/ivoireai/suggestedQueries.ts
@@ -1,0 +1,8 @@
+  const suggestedQueries = [
+    "Combien d'apprenants sont inscrits dans la région d'Abidjan ?",
+    "Montre-moi la répartition des centres par région",
+    "Quelles sont les tendances d'inscription sur les 6 derniers mois ?",
+    "Génère un rapport sur les taux de réussite par filière",
+  ];
+
+  export default suggestedQueries
