+++ frontend/components/ivoireai/tabsList.ts
@@ -1,0 +1,21 @@
+import { Bot, Database, FileText } from "lucide-react";
+
+const tabsList = [
+    {
+      value: "chat",
+      icon: Bot,
+      label: "Chat",
+    },
+    {
+      value: "data",
+      icon: Database,
+      label: "Données",
+    },
+    {
+      value: "docs",
+      icon: FileText,
+      label: "Documents",
+    },
+  ];
+  
+  export default tabsList
